% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la.vax}
\alias{la.vax}
\title{Vaccination rates by CSA and date in Los Angeles County}
\format{
TODO: complete this
}
\source{
\url{http://publichealth.lacounty.gov/media/coronavirus/vaccine/vaccine-dashboard.htm#}
}
\usage{
la.vax
}
\description{
a dataset containing vaccination rates by CSA and date in Los Angeles County
}
\keyword{datasets}
