% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Reformat data}
\usage{
load_data(
  data_folder_path = fs::path("inst/extdata", "LAC_Vaccine_City_Data.csv")
)
}
\arguments{
\item{data_folder_path}{path to the csv file downloaded from LACDPH vaccination website}
}
\value{
a tibble containing the variables of interset
}
\description{
Reformat data
}
